Refactor: api route 변경 후, 적용
